Modify page initialization to pull data from TagCollection
